requirements.txt
asyncio==3.4.3
awesome-slugify==1.6.5
certifi==2018.11.29
chardet==3.0.4
Django==1.11
django-mathfilters==0.4.0
docutils==0.14
idna==2.8
lockfile==0.12.2
Pillow==5.4.1
psycopg2==2.7.6.1
pycrypto==2.6.1
python-daemon==2.2.3
python-resize-image==1.1.18
pytz==2018.9
regex==2019.2.7
requests==2.21.0
six==1.12.0
Unidecode==0.4.21
urllib3==1.24.1
